Eliminated dags from file bshrubs/btrees
Velociraptors inbound.

This eliminates dags (directed acyclic graphs) from file bshrubs/btrees,
which were the only source of dags in the filesystem. This means
littlefs is now strictly a pure tree, in that no blocks have more than
one parent (ignoring in-RAM references!).

Up until this point, dags could be created in file bshrubs/btrees via
random writes that place fragments in the middle of a block:

  .-------------.      .-------------------.
  | aaaaaaaaaaa |  ->  | aaaaa | b | aaaaa |
  '-------------'      '-------------------'
         |                  |    v    |
         v                  |   .-.   |
  .-------------.           |   |b|   |
  | aaaaaaaaaaa |           v   '-'   v
  '-------------'         .-------------.
                          | aaaaaaaaaaa |
                          '-------------'

Now, fragments that would create dags instead trigger block
recrystallization, rewriting the left sibling into a new block if
necessary:

  .-------------.      .----------------.
  | aaaaaaaaaaa |  ->  | aaaaab | aaaaa |
  '-------------'      '----------------'
         |                 |        '-.
         v                 v          v
  .-------------.      .--------. .-------.
  | aaaaaaaaaaa |      | aaaaab | | aaaaa |
  '-------------'      '--------' '-------'

Allowing dags was great for random-write performance, but it creates
problems for future planned features:

1. Current plans for more advanced block allocators rely on blocks only
   having one parent. Otherwise it's difficult to know which reference
   is the last reference to a block.

2. Dags create a really funny problem for error correction via block
   redundancy. Naively, if you try to repair blocks every time you
   encounter a given block error, you will end up exploding the block
   into n copies, 1 for every parent. Not great!

---

Eliminating these dags was a bit... tricky...

Originally I was planning to just alloc/rewrite blocks in
lfsr_file_carve, but it turns out we can make lfsr_file_flush_ do all
the work with an extra would-dag checks. Handling dags in
lfsr_file_flush_ also gives us a chance to merge any pending data and
get the most out of the block rewrite.

This does give us a bit of technical debt in that we will probably still
need the block splitting in lfsr_file_carve for future features
(advanced hole APIs, alternative write strategies, etc), but it's
probably worth it for code savings in the default build.

Unfortunately this does add to the mess that is lfsr_file_flush_'s
control flow graph:

         lfsr_file_flush_
               |
               v
  .--> lookup left crystal   .--> lookup left sibling <-.
  |            |             |            |             |
  |            v             |            v             |
  |         erased?          |           dag? (new!)    |
  | .---------y n            | .---------y n            |
  | |           v            | |           v            |
  | |  lookup right crystal  | |  lookup right sibling  |
  | |          |             | |          |             |
  | |          v             | |          v             |
  | |   >=crystal_thresh?    | |       coalesce         |
  | |         y n------------' |          |             |
  | |         v                |          v             |
  | |  lookup left neighbor    |        carve-----------'
  | |          |               |
  | |          v               |
  | |       erased?            |
  | +---------y n              |
  | |           v              |
  | |        alloc <---+-------'
  | |          |       |
  | |          v       |
  | '---> crystallize  |
  |            |       |
  |            v       |
  |          good?     |
  |           y n------'
  |           v
  '----------carve

I did scratch my head for a bit trying to think if there was a better
way to organize this, but came up empty.

It looks complicated, but we really only have two* loops (ignoring the
relocation loop): One that crystallizes blocks, and one that coalesces
fragments. The problem is that we end jumping between the two depending
on what we find in the btree.

In a sane system, this would be implemented as mutually recursive
functions, but this is littlefs, the whole point is that we don't use
recursion.

---

The good news is that this added surprisingly little code (and saved
stack?):

           code          stack          ctx
  before: 35600           2448          640
  after:  35692 (+0.3%)   2432 (-0.7%)  640 (+0.0%)