Another ckparity flaw
Found another ckparity flaw! Only detected now due to the reworked tag
encoding, but it's just luck this wasn't detected earlier.

Consider the following bit flip:

  03 04 80 04 80 04 6b ...  data w512 512
  43 04 80 04 80 04 6b ...  altble 0x304 w512 -512
  ^
  flip

Not only are leb128s problem for ckparity, but even the difference in
alt vs normal tag encoding presents a vulnerability.

So limiting the ckparity tests further, to just 47 of the first 48 bits.